Reorganized current test functions a bit
